ml/pyproject.toml
[project]
name = "ecosort-ml"
version = "0.1.0"
description = "Training and export pipeline for EcoSort mobile waste classification."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "Pillow>=10.3",
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "matplotlib>=3.9",
  "seaborn>=0.13",
  "tqdm>=4.66"
]

[tool.setuptools.packages.find]
where = ["src"]

